requirements.txt
Flask==3.1.2
flask-cors==6.0.1
tensorflow==2.20.0
keras==3.11.3
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.7.2
scipy==1.16.2
matplotlib==3.10.7
pillow==12.0.0
tqdm==4.67.1
protobuf==5.29.5
requests==2.32.5
google-auth==2.42.0
google-generativeai==0.8.5
joblib==1.5.2
rich==14.2.0
Werkzeug==3.1.3
Jinja2==3.1.6
colorama==0.4.6
typing_extensions==4.15.0
packaging==25.0
